GalaxeaDP/requirements.txt
sharedarray==3.2.4
yapf==0.40.2
addict==2.4.0
plyfile==1.0.3
termcolor==2.4.0
hydra-colorlog==1.2.0
hydra-core==1.3.2
hydra-optuna-sweeper==1.2.0
pre-commit==3.7.1
pytest==8.2.2
rich==13.7.1
isort==5.13.2
black==24.4.2
rootutils==1.0.7
python-dotenv==1.0.1
gdown==5.2.0
ipdb==0.13.13
h5py==3.14.0
h5py-cache==1.0.0
zarr==2.18.2
matplotlib==3.10.3
plotly==6.2.0
toml==0.10.2
wandb==0.21.0
loguru==0.7.3
einops==0.8.0
numpy==1.26.4
open3d==0.18.0
opencv-python==4.11.0.86
scipy==1.15.3
transforms3d==0.4.1
imageio==2.34.1
diffusers==0.29.0
datasets==3.6.0
huggingface-hub==0.33.4
timm==0.9.10
gymnasium==0.29.1
torch==2.7.1
lightning==2.5.2
torchmetrics==1.7.4
torch-geometric==2.5.3
torchcodec==0.4.0
sapien==3.0.0b1
trimesh==4.7.3
coacd==1.0.5
tyro==0.9.28
pytorch_kinematics_ms==0.7.3
seaborn==0.13.2
jsonlines==4.0.0
draccus==0.10.0
av==15.0.0
deepdiff==8.5.0
numba==0.61.2
pymunk==6.11.1
pyserial==3.5
https://github.com/haosulab/MPlib/releases/download/nightly/mplib-0.2.0.dev20241021+git.da9e0614-cp310-cp310-manylinux_2_17_x86_64.manylinux2014_x86_64.whl